Conditionals for dev & prod env. Sign S3 files before delivering to the client

+++ server/controllers/technology.controller.js
@@ -1,6 +1,6 @@
 const path = require('path')
 
-const { s3Upload, s3Delete } = require('../config/aws.config')
+const { s3UploadFile, s3DeleteFile, s3GetSignedURL } = require('../config/aws.config')
 const Technology = require('../models/Technology')
 const StatusCodeError = require('../utils/statusCodeError')
 
@@ -9,15 +9,20 @@ const techCreate = async (req, res) => {
   const userId = req.userId
   const image = req.file
 
-  if (!(name && image && userId))
-    throw new StatusCodeError(400, 'Your new technology should include the name and logo.')
+  if (!userId) throw new StatusCodeError(401, 'The user is not authenticated to make this request.')
+
+  if (!(name && image))
+    throw new StatusCodeError(400, 'A new technology should include the name and logo.')
 
   try {
-    var imageURL = await s3Upload(name + path.extname(image.originalname), image.path, 'technologies')
+    var imageURL = await s3UploadFile(name + path.extname(image.originalname), image.path, 'technologies')
     
     const technology = new Technology({ name, docsURL, imageURL, userId })
     
     await technology.save()
+
+    technology.imageURL = await s3GetSignedURL(imageURL)
+    
     return res.status(200).json({
       alert: {
         type: 'success',
@@ -26,7 +31,7 @@ const techCreate = async (req, res) => {
       technology
     })
   } catch (error) {
-    if (imageURL) s3Delete(imageURL)
+    if (imageURL) s3DeleteFile(imageURL)
     console.error(error)
     return res.status(error?.statusCode || 400).json({
       alert: {
@@ -42,11 +47,15 @@ const techDeleteOne = async (req, res) => {
   const userId = req.userId
   
   try {
+    if (!userId) throw new StatusCodeError(401, 'The user is not authenticated to make this request.')
+    
+    if (!techId) throw new StatusCodeError(404, 'An ID for the technology to be deleted was not provided.')
+
     const technology = await Technology.findOneAndDelete({ _id: techId, userId })
 
     if (!technology) throw new StatusCodeError(404, 'Could not find a technology that matches the technology ID and user ID.')
     
-    const techImage = await s3Delete(technology.imageURL)
+    const techImage = await s3DeleteFile(technology.imageURL)
     if (!techImage) throw new Error('Image was not removed from the cloud.')
 
     return res.status(200).json({
